Move Loose: Spectral Library Tests
