+++ mzLib/Test/FileReadingTests/SpectralLibraryTests/SpectralLibraryData/myPrositLib.msp
